requirements.txt
tensorflow>=2.9
google-cloud-storage>=2.0
pandas>=1.3
scikit-learn>=1.0
pyspark>=3.3

